Removed old and duplicate files
There were several example of files that were checked in as separate
from the main and the old tag appended.  Since we are using version
control a file's history is part of the repo and doesn't need to be
stored in this way.  Those files are removed now!

There were also two copies of our libraries (in /js and /js/lib)
and now they only exist in the proper /js/lib direcory.